refactor: scan-job reports
