+++ pkg/vulnerabilityreport/io.go
@@ -2,11 +2,17 @@ package vulnerabilityreport
 
 import (
 	"context"
+	"time"
 
+	"github.com/aquasecurity/trivy-db/pkg/types"
 	"github.com/aquasecurity/trivy-operator/pkg/apis/aquasecurity/v1alpha1"
+	"github.com/aquasecurity/trivy-operator/pkg/ext"
 	"github.com/aquasecurity/trivy-operator/pkg/kube"
+	ty "github.com/aquasecurity/trivy/pkg/types"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/types"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	tn "k8s.io/apimachinery/pkg/types"
+	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -56,7 +62,7 @@ func (r *readWriter) Write(ctx context.Context, reports []v1alpha1.Vulnerability
 
 func (r *readWriter) createOrUpdate(ctx context.Context, report v1alpha1.VulnerabilityReport) error {
 	var existing v1alpha1.VulnerabilityReport
-	err := r.Get(ctx, types.NamespacedName{
+	err := r.Get(ctx, tn.NamespacedName{
 		Name:      report.Name,
 		Namespace: report.Namespace,
 	}, &existing)
@@ -87,4 +93,135 @@ func (r *readWriter) FindByOwner(ctx context.Context, owner kube.ObjectRef) ([]v
 	}
 
 	return list.DeepCopy().Items, nil
+}
+
+type AdditionalFields struct {
+	Description bool
+	Links       bool
+	CVSS        bool
+	Target      bool
+	Class       bool
+	PackageType bool
+	PkgPath     bool
+}
+
+func GetVulnerabilitiesFromScanResult(report ty.Result, addFields AdditionalFields) []v1alpha1.Vulnerability {
+	vulnerabilities := make([]v1alpha1.Vulnerability, 0)
+
+	for _, sr := range report.Vulnerabilities {
+		var pd, lmd string
+		if sr.PublishedDate != nil {
+			pd = sr.PublishedDate.Format(time.RFC3339)
+		}
+		if sr.LastModifiedDate != nil {
+			lmd = sr.LastModifiedDate.Format(time.RFC3339)
+		}
+		vulnerability := v1alpha1.Vulnerability{
+			VulnerabilityID:  sr.VulnerabilityID,
+			Resource:         sr.PkgName,
+			InstalledVersion: sr.InstalledVersion,
+			FixedVersion:     sr.FixedVersion,
+			PublishedDate:    pd,
+			LastModifiedDate: lmd,
+			Severity:         v1alpha1.Severity(sr.Severity),
+			Title:            sr.Title,
+			PrimaryLink:      sr.PrimaryURL,
+			Links:            []string{},
+			Score:            GetScoreFromCVSS(GetCvssV3(sr.CVSS)),
+		}
+
+		if addFields.Description {
+			vulnerability.Description = sr.Description
+		}
+		if addFields.Links && sr.References != nil {
+			vulnerability.Links = sr.References
+		}
+		if addFields.CVSS {
+			vulnerability.CVSS = sr.CVSS
+		}
+		if addFields.Target {
+			vulnerability.Target = report.Target
+		}
+		if addFields.Class {
+			vulnerability.Class = string(report.Class)
+		}
+		if addFields.PackageType {
+			vulnerability.PackageType = string(report.Type)
+		}
+		if addFields.PkgPath {
+			vulnerability.PkgPath = sr.PkgPath
+		}
+
+		vulnerabilities = append(vulnerabilities, vulnerability)
+	}
+
+	return vulnerabilities
+}
+
+func GetCvssV3(findingCvss types.VendorCVSS) map[string]*CVSS {
+	cvssV3 := make(map[string]*CVSS)
+	for vendor, cvss := range findingCvss {
+		var v3Score *float64
+		if cvss.V3Score != 0.0 {
+			v3Score = ptr.To[float64](cvss.V3Score)
+		}
+		cvssV3[string(vendor)] = &CVSS{v3Score}
+	}
+	return cvssV3
+}
+
+func GetScoreFromCVSS(CVSSs map[string]*CVSS) *float64 {
+	var nvdScore, vendorScore *float64
+
+	for name, cvss := range CVSSs {
+		if name == "nvd" {
+			nvdScore = cvss.V3Score
+		} else {
+			vendorScore = cvss.V3Score
+		}
+	}
+
+	if nvdScore != nil {
+		return nvdScore
+	}
+
+	return vendorScore
+}
+
+type CVSS struct {
+	V3Score *float64 `json:"V3Score,omitempty"`
+}
+
+func BuildVulnerabilityReportData(clock ext.Clock, registry v1alpha1.Registry, artifact v1alpha1.Artifact, version string, vulnerabilities []v1alpha1.Vulnerability) v1alpha1.VulnerabilityReportData {
+	return v1alpha1.VulnerabilityReportData{
+		UpdateTimestamp: metav1.NewTime(clock.Now()),
+		Scanner: v1alpha1.Scanner{
+			Name:    v1alpha1.ScannerNameTrivy,
+			Vendor:  "Aqua Security",
+			Version: version,
+		},
+		Registry:        registry,
+		Artifact:        artifact,
+		Summary:         vulnerabilitySummary(vulnerabilities),
+		Vulnerabilities: vulnerabilities,
+	}
+}
+
+func vulnerabilitySummary(vulnerabilities []v1alpha1.Vulnerability) v1alpha1.VulnerabilitySummary {
+	var vs v1alpha1.VulnerabilitySummary
+	for _, v := range vulnerabilities {
+		switch v.Severity {
+		case v1alpha1.SeverityCritical:
+			vs.CriticalCount++
+		case v1alpha1.SeverityHigh:
+			vs.HighCount++
+		case v1alpha1.SeverityMedium:
+			vs.MediumCount++
+		case v1alpha1.SeverityLow:
+			vs.LowCount++
+		default:
+			vs.UnknownCount++
+		}
+	}
+	return vs
 }
